requirements.txt
# requirements.txt
tensorflow>=2.12.0       # includes tf.keras
numpy>=1.24.0
pillow>=9.5.0